% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{Collection}
\alias{Collection}
\alias{Collection,character,data.frame,character,character-method}
\alias{Collection,character,data.frame,missing,missing-method}
\alias{Collection,character,character,missing,missing-method}
\alias{Collection,character,character,character,character-method}
\alias{Collection,missing,missing,missing,missing-method}
\title{Create a Collection}
\usage{
Collection(name, data, recordIdColumn, ancestorIdColumns)

\S4method{Collection}{character,data.frame,character,character}(name, data, recordIdColumn, ancestorIdColumns)

\S4method{Collection}{character,data.frame,missing,missing}(name, data, recordIdColumn, ancestorIdColumns)

\S4method{Collection}{character,character,missing,missing}(name, data, recordIdColumn, ancestorIdColumns)

\S4method{Collection}{character,character,character,character}(name, data, recordIdColumn, ancestorIdColumns)

\S4method{Collection}{missing,missing,missing,missing}(name, data, recordIdColumn, ancestorIdColumns)
}
\arguments{
\item{name}{The name of the collection}

\item{data}{A data.frame or a character vector representing a file path to a data.frame}

\item{recordIdColumn}{The name of the column in the data.frame that contains the record id}
}
\description{
This is a constructor for the Collection class. It creates a Collection
object for use in a Microbiome Dataset.
}
