% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{Collection}
\alias{Collection}
\title{Create a Collection}
\usage{
Collection(name, data, recordIdColumn)
}
\arguments{
\item{name}{The name of the collection}

\item{data}{A data.frame or a character vector representing a file path to a data.frame}

\item{recordIdColumn}{The name of the column in the data.frame that contains the record id}
}
\description{
This is a constructor for the Collection class. It creates a Collection
object for use in a Microbiome Dataset.
}
