% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{Collections}
\alias{Collections}
\alias{Collections,missing,missing-method}
\alias{Collections,list,missing-method}
\alias{Collections,list,data.frame-method}
\alias{Collections,data.frame,missing-method}
\alias{Collections,data.frame,data.frame-method}
\alias{Collections,Collection,missing-method}
\alias{Collections,character,missing-method}
\title{Create Collections}
\usage{
Collections(collections, ontology)

\S4method{Collections}{missing,missing}(collections, ontology)

\S4method{Collections}{list,missing}(collections, ontology)

\S4method{Collections}{list,data.frame}(collections, ontology)

\S4method{Collections}{data.frame,missing}(collections, ontology)

\S4method{Collections}{data.frame,data.frame}(collections, ontology)

\S4method{Collections}{Collection,missing}(collections, ontology)

\S4method{Collections}{character,missing}(collections, ontology)
}
\arguments{
\item{collections}{A list of Collection objects, a data.frame containing multiple
collections, or a character vector containing a file path to a data.frame}

\item{ontology}{A data.frame containing the ontology for the dataset}
}
\value{
A Collections object
}
\description{
This is a constructor for the Collections class. It creates a Collections
object for use in a Microbiome Dataset. A Collections object is a list of
Collection objects.
}
