% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DiabImmune}
\alias{DiabImmune}
\title{DiabImmune}
\format{
\subsection{\code{DiabImmune}}{

A MbioDataset object with 60 metadata variables, 3184 16S stool samples and 1149 shotgun stool samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
"Shotgun metagenomics: 4th level EC metagenome abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway coverage data"  \cr
"Shotgun metagenomics: Genus" \cr
"Shotgun metagenomics: Species" \cr
"Shotgun metagenomics: Family" \cr
"Shotgun metagenomics: Order" \cr
"Shotgun metagenomics: Phylum" \cr
"Shotgun metagenomics: Class"  \cr
"Shotgun metagenomics: Kingdom"  \cr
"Shotgun metagenomics: Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_a2f8877e68/new/download}
}
\usage{
DiabImmune
}
\description{
Broad Institute's study of three pediatric cohorts that examined the microbiome in
children at high risk for development of Type 1 diabetes (T1D).
Includes 289 children across three sites, including Russia (Karelia), Finland and Estonia.
There are 3184 stool samples which have amplicon sequencing data from the V4 region of 16S rRNA gene.
There are 1149 stool samples having 'shotgun' metagenomic sequencing data.
Prospective cohort design with monthly sampling for the first ~3 years of life.
Website Release 32. (2023 May 30)
}
\keyword{datasets}
