% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FARMM}
\alias{FARMM}
\title{FARMM}
\format{
\subsection{\code{FARMM}}{

A MbioDataset object with 17 metadata variables, 454 shotgun stool assays and 150 mass spec assays.
It contains the following collections: \cr
"Metabolomics Mass spectrometry assay"
"Shotgun metagenomics: 4th level EC metagenome abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway coverage data"  \cr
"Shotgun metagenomics: Genus" \cr
"Shotgun metagenomics: Species" \cr
"Shotgun metagenomics: Family" \cr
"Shotgun metagenomics: Order" \cr
"Shotgun metagenomics: Phylum" \cr
"Shotgun metagenomics: Class"  \cr
"Shotgun metagenomics: Kingdom"  \cr
"Shotgun metagenomics: Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_4dfda49064/new/download}
}
\usage{
FARMM
}
\description{
The Food and Resulting Microbial Metabolites (FARMM) study set out to define the impact of defined diets on the microbiome and metabolome.
Follows 31 healthy human adult volunteers monitored longitudinally over a 15 day period.
Includes 454 stool samples and sequencing controls; 'shotgun' metagenomic sequencing.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
