% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEMS1}
\alias{GEMS1}
\title{GEMS1 Case Control}
\format{
\subsection{\code{GEMS1}}{

A MbioDataset object with 12 metadata variables, 1015 16S stool samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_a5167b81e3/new/download}
}
\usage{
GEMS1
}
\description{
The Global Enteric Multicenter Study (GEMS) investigated the cause, incidence and impact of
moderate to severe diarrheal (MSD) disease in children. 514 MSD cases and 493 endemic controls from Bangladesh,
The Gambia, Kenya and Mali were included. 1007 stool samples; V1-V2 region of 16S rRNA gene.
Prospective, age-stratified, matched case-control design. Additional clinical and epidemiologic
data for the same participants is available at ClinEpiDB.org.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
