% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollection}
\alias{getCollection}
\title{Get Microbiome Dataset Collection}
\usage{
getCollection(object, collectionName)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{collectionName}{The name of the collection to return}
}
\value{
An AbundanceData object representing the collection and any associated study metadata
}
\description{
Get a collection from the Microbiome Dataset. The collection will be returned
as an AbundanceData object.
}
