% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollectionNames}
\alias{getCollectionNames}
\title{Get Microbiome Dataset Collection Names}
\usage{
getCollectionNames(object)
}
\arguments{
\item{object}{A Microbiome Dataset}
}
\value{
a character vector of collection names
}
\description{
Get the names of the collections in the Microbiome Dataset.
}
