% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getCuratedDatasetNames}
\alias{getCuratedDatasetNames}
\title{Get Curated Data Set Names}
\usage{
getCuratedDatasetNames()
}
\value{
A vector of curated data set names
}
\description{
Returns a vector of curated data set names from the 'microbiomeData' package.
These data are the same from the MicrobiomeDB.org website and can be
analyzed with the 'MicrobiomeDB' R package.
}
