% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getIdColumnNames}
\alias{getIdColumnNames}
\title{Get Microbiome Dataset Id Column Names}
\usage{
getIdColumnNames(object)
}
\arguments{
\item{object}{A Microbiome Dataset, or other object w recordIdColumn and ancestorIdColumns slots}
}
\value{
a character vector of id column names
}
\description{
Get the names of the record and ancestor id columns in the Microbiome Dataset.
}
