% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R,
%   R/methods-Collections.R, R/methods-MbioDataset.R
\name{getMetadataVariableNames}
\alias{getMetadataVariableNames}
\alias{getMetadataVariableNames,AbundanceData-method}
\alias{getMetadataVariableNames,Collection-method}
\alias{getMetadataVariableNames,MbioDataset-method}
\title{Get Microbiome Dataset Metadata Variable Names}
\usage{
getMetadataVariableNames(object)

\S4method{getMetadataVariableNames}{AbundanceData}(object)

\S4method{getMetadataVariableNames}{Collection}(object)

\S4method{getMetadataVariableNames}{MbioDataset}(object)
}
\arguments{
\item{object}{A Microbiome Dataset}
}
\value{
a character vector of metadata variable names
}
\description{
Get the names of the metadata variables in the Microbiome Dataset.
}
