% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getSampleMetadata}
\alias{getSampleMetadata}
\title{Get Microbiome Dataset Sample Metadata}
\usage{
getSampleMetadata(object, metadataVariables = NULL)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{metadataVariables}{The metadata variables to include in the sample metadata. If NULL, all metadata variables will be included.}
}
\value{
A data.table of sample metadata
}
\description{
Get the sample metadata from the Microbiome Dataset. If metadata variables are not specified, all sample metadata will be returned.
}
