% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP_MGX}
\alias{HMP_MGX}
\title{HMP Phase 1 Shotgun Metagenomics}
\format{
\subsection{\code{HMP_MGX}}{

A MbioDataset object with 14 metadata variables, 2145 shotgun metagenomic stool samples.
It contains the following collections: \cr
"Shotgun metagenomics: 4th level EC metagenome abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway coverage data"  \cr
"Shotgun metagenomics: Genus" \cr
"Shotgun metagenomics: Species" \cr
"Shotgun metagenomics: Family" \cr
"Shotgun metagenomics: Order" \cr
"Shotgun metagenomics: Phylum" \cr
"Shotgun metagenomics: Class"  \cr
"Shotgun metagenomics: Kingdom"  \cr
"Shotgun metagenomics: Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_898df5869d/new/download}
}
\usage{
HMP_MGX
}
\description{
The Human Microbiome Project (HMP) profiled microbial communities across diverse habitats on the human body.
103 healthly adult volunteers sampled at up to 20 different body sites.
741 samples of various types, including stool, saliva, and oral, vaginal and nasal swabs; 'shotgun' metagenomic sequencing.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
