% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP_V3V5}
\alias{HMP_V3V5}
\title{HMP Phase 1 (V3-V5)}
\format{
\subsection{\code{HMP_V3V5}}{

A MbioDataset object with 13 metadata variables, 5462 16S samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_ca4404e155/new/download}
}
\usage{
HMP_V3V5
}
\description{
The Human Microbiome Project (HMP) profiled microbial communities across diverse habitats on the human body.
242 healthly adult volunteers sampled at approximately 15 different body sites.
5826 samples of various types; V3-V5 region of 16S rRNA gene.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
