% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP_WGS}
\alias{HMP_WGS}
\title{HMP Phase 1 WGS}
\format{
\subsection{\code{HMP_WGS}}{

A MbioDataset object with 14 metadata variables, 2145 WGS stool samples.
It contains the following collections: \cr
"WGS 4th level EC metagenome abundance data" \cr
"WGS Metagenome enzyme pathway abundance data" \cr
"WGS Metagenome enzyme pathway coverage data"  \cr
"WGS Genus" \cr
"WGS Species" \cr
"WGS Family" \cr
"WGS Order" \cr
"WGS Phylum" \cr
"WGS Class"  \cr
"WGS Kingdom"  \cr
"WGS Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_898df5869d/new/download}
}
\usage{
HMP_WGS
}
\description{
The Human Microbiome Project (HMP) profiled microbial communities across diverse habitats on the human body.
103 healthly adult volunteers sampled at up to 20 different body sites.
741 samples of various types, including stool, saliva, and oral, vaginal and nasal swabs; 'shotgun' metagenomic sequencing.
}
\keyword{datasets}
