% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leishmaniasis}
\alias{Leishmaniasis}
\title{Human cutaneous leishmaniasis}
\format{
\subsection{\code{Leishmaniasis}}{

A MbioDataset object with 16 metadata variables, 128 16S samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_75221fce95/new/download}
}
\usage{
Leishmaniasis
}
\description{
This study explored the skin microbiome during human cutaneous leishmaniasis.
45 patients with cutaneous leishmaniasis.
128 skin swabs collected from affected and unaffected skin; V1-V3 region of 16S rRNA gene.
Case-control (restrospective) design in patients seen at a field hospital
(Clinica em Leishmaniose Dr. Jackson Costa) in Corte de Pedra, Bahia, Brazil.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
