% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MALED_2yr}
\alias{MALED_2yr}
\title{MAL-ED 0-2yr}
\format{
\subsection{\code{MALED_2yr}}{

A MbioDataset object with 16 metadata variables, 1091 16S stool samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_72c94486c6/new/download}
}
\usage{
MALED_2yr
}
\description{
This study set out to define the normal maturation of the gut microbiome during the first 2 years of postnatal life.
50 children from four sites in the MAL-ED study: Brazil, India, Peru and South Africa.
1091 stool samples; V4 region of 16S rRNA gene.
Prospective cohort design with monthly sampling for the first ~2 years of life.
Additional clinical and epidemiologic data for the same participants is available at ClinEpiDB.org.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
