% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{MbioDataset,data.frame,character-method}
\alias{MbioDataset,data.frame,character-method}
\title{export}
\usage{
\S4method{MbioDataset}{data.frame,character}(collections, metadata)
}
\description{
export
}
