% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{MbioDataset}
\alias{MbioDataset}
\title{Create a Microbiome Dataset}
\usage{
MbioDataset(collections, metadata, ontology)
}
\arguments{
\item{collections}{A list of Collection objects, a data.frame containing multiple collections,
or a character vector containing one or more file path(s)}

\item{metadata}{A SampleMetadata object, a data.frame containing sample metadata,
or a list of file path(s)}

\item{ontology}{An data.frame containing the ontology of the dataset, or a character vector
containing a file path to a data.frame}
}
\description{
This is a constructor for the MbioDataset class. It creates a MbioDataset containing
a list of Collections and a SampleMetadata object.
}
