% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeData-package.R
\docType{package}
\name{microbiomeData-package}
\alias{microbiomeData}
\alias{microbiomeData-package}
\title{microbiomeData: Curated Data from MicrobiomeDB.org}
\description{
An R package containing all of the curated datasets from MicrobiomeDB.org.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/microbiomeDB/microbiomeData}
  \item \url{https://microbiomedb.github.io/microbiomeData/}
  \item Report bugs at \url{https://github.com/microbiomeDB/microbiomeData/issues}
}

}
\author{
\strong{Maintainer}: Danielle Callan \email{dcallan@upenn.edu}

}
\keyword{internal}
