% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MORDOR}
\alias{MORDOR}
\title{MORDOR phase 1}
\format{
\subsection{\code{MORDOR}}{

A MbioDataset object with 10 metadata variables, 574 shotgun stool samples.
It contains the following collections: \cr
"Shotgun metagenomics: 4th level EC metagenome abundance data" \cr
"Shotgun metagenomics: Metagenome enzyme pathway abundance data" \cr
"Shotgun metagenomics: Metagenomeenzyme pathway coverage data"  \cr
"Shotgun metagenomics: Genus" \cr
"Shotgun metagenomics: Species" \cr
"Shotgun metagenomics: Family" \cr
"Shotgun metagenomics: Order" \cr
"Shotgun metagenomics: Phylum" \cr
"Shotgun metagenomics: Class"  \cr
"Shotgun metagenomics: Kingdom"  \cr
"Shotgun metagenomics: Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_39d3629e79/new/download}
}
\usage{
MORDOR
}
\description{
The Macrolides Oraux pour Réduire les Décès avec un Oeil sur la Résistance (MORDOR) I
metagenome study evaluated the impact of mass azithromycin administration on the gut microbiome in children
574 pediatric (preschool) participants from 30 Nigerien communities
Rectal swabs; 'shotgun' metagenomic sequencing of RNA.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
