% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{name}
\alias{name}
\alias{name,Collection-method}
\alias{name<-}
\alias{name<-,Collection-method}
\alias{name<-,Collection,character-method}
\title{Manage Collection Names}
\usage{
name(x)

\S4method{name}{Collection}(x)

name(x) <- value

\S4method{name}{Collection}(x) <- value
}
\arguments{
\item{x}{A Collection}
}
\description{
Get or set the name of the collection
}
