% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NICU_NEC}
\alias{NICU_NEC}
\title{NICU-NEC}
\format{
\subsection{\code{NICU_NEC}}{

A MbioDataset object with 46 metadata variables, 1118 WGS stool samples.
It contains the following collections: \cr
"WGS 4th level EC metagenome abundance data" \cr
"WGS Metagenome enzyme pathway abundance data" \cr
"WGS Metagenome enzyme pathway coverage data"  \cr
"WGS Genus" \cr
"WGS Species" \cr
"WGS Family" \cr
"WGS Order" \cr
"WGS Phylum" \cr
"WGS Class"  \cr
"WGS Kingdom"  \cr
"WGS Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_84fcb69f4e/new/download}
}
\usage{
NICU_NEC()
}
\description{
The Neonatal Intensive Care Unit, Necrotizing Enterocolitis (NICU NEC) study set out to understand microbial factors associated with NEC onset
150 infants born prematurely were sampled perinatally and at high frequency during the first few weeks to months of life
Prospective cohort design.
}
