% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PIH_Uganda}
\alias{PIH_Uganda}
\title{PIH Uganda}
\format{
\subsection{\code{PIH_Uganda}}{

A MbioDataset object with 11 metadata variables, 92 CSF samples.
It contains the following collections: \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_d00b1afb83/new/download}
}
\usage{
PIH_Uganda
}
\description{
This study profiled microbes present in the cerebrospinal fluid (CSF) of pediatric patients
with infectious and non-infectious hydrocephalus. 92 children from Uganda, including 58 with
postinfectious hydrocephalus (PIH) and 34 with non-postinfectious hydrocephalus (NPIH).
92 cerebrospinal fluid (CSF) samples retained in this workflow; V1-V2 region of 16S rRNA gene.
Case-control (retrospective) design with samples collected in the first ~100 days of life.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
