% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PretermInfantResistome1}
\alias{PretermInfantResistome1}
\title{Preterm Infant Resistome 1}
\format{
\subsection{\code{PretermInfantResistome1}}{

A MbioDataset object with 64 metadata variables, 399 shotgun stool samples, 103 16S stool samples.
It contains the following collections: \cr
"Shotgun metagenomics: 4th level EC metagenome abundance data" \cr
"Shotgun metagenomics: Metagenomeenzyme pathway abundance data" \cr
"Shotgun metagenomics: Metagenomeenzyme pathway coverage data"  \cr
"Shotgun metagenomics: Genus" \cr
"Shotgun metagenomics: Species" \cr
"Shotgun metagenomics: Family" \cr
"Shotgun metagenomics: Order" \cr
"Shotgun metagenomics: Phylum" \cr
"Shotgun metagenomics: Class"  \cr
"Shotgun metagenomics: Kingdom"  \cr
"Shotgun metagenomics: Normalized number of taxon-specific sequence matches" \cr
"16S Order" \cr
"16S Genus" \cr
"16S Family" \cr
"16S Species" \cr
"16S Class" \cr
"16S Phylum" \cr
"16S Kingdom" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_d1b9f788dc/new/download}
}
\usage{
PretermInfantResistome1
}
\description{
The Preterm Infant Resistome I study used 16S rRNA marker gene sequencing and shotgun
metagenomics to profile the stool metagenome in preterm infants (born < 33 weeks gestational age)
stratified by antibiotic use during their hospitalization.
84 preterm infant patients seen at the Washington University School of Medicine in St. Louis, MO, USA.
399 stool samples; 'shotgun' metagenomic sequencing.
103/399 stool samples; V4 16S rRNA gene sequencing.
Prospective cohort design.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
