% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rats_inulin}
\alias{rats_inulin}
\title{Microbiome of fecal from rats fed with sausage fortified with Inulin}
\format{
A phyloseq object with 480 features and 59 samples
\describe{
\item{time}{before (Start) or after (slut) dietary intervention}
\item{Description}{Type of dietary intervention}
\item{ID}{Rat ID}
\item{Acetate to Xylose}{Associated Molecular markers}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/mnfr.201800608?casa_token=IEsQ-Ry1wQUAAAAA\%3AsbD6qevecfM9lUXvc7YnU0SE6sGN4lgHROsjPshSRcoWS0S1mJG060eDZX4yfme8TZJzLNr6MbhhVs0}
}
\usage{
rats_inulin
}
\description{
A phyloseq object containing 16s sequencing of fecal samples from rats feed different diets
}
\examples{
phyloseq::sample_sums(rats_inulin)
}
\references{
Thøgersen, Rebekka, et al. "Ingestion of an Inulin‐Enriched Pork Sausage Product Positively Modulates the Gut Microbiome and Metabolome of Healthy Rats." Molecular nutrition & food research 62.19 (2018): 1800608.
}
\keyword{datasets}
