% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{testOTU}
\alias{testOTU}
\title{testOTU}
\format{
\subsection{\code{testOTU}}{

A data.table with hundreds of mock OTUs
}
}
\usage{
testOTU()
}
\description{
A test dataset for testing the functions in the package
}
