% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{updateCollectionName}
\alias{updateCollectionName}
\title{Update Microbiome Dataset Collection Name}
\usage{
updateCollectionName(object, oldName, newName)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{oldName}{The name of the collection to update}

\item{newName}{The new name of the collection}
}
\value{
A Microbiome Dataset with the updated collection name
}
\description{
Update the name of a collection in the Microbiome Dataset.
}
