% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MbioDataset.R
\docType{class}
\name{Collection-class}
\alias{Collection-class}
\title{Microbiome Data Collection}
\description{
#' This class represents a collection of Microbiome Data. That could be
all of the abundance data for all samples in a dataset at a particular taxonomic rank,
or all pathway abundance data for all samples in a dataset, or something else. The
primary requirement for a collection is that the values of all variables in the
collection can be expressed on the same theoretical range.
}
