% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{Collections}
\alias{Collections}
\title{Create Collections}
\usage{
Collections(collections)
}
\arguments{
\item{collections}{A list of Collection objects, a data.frame containing multiple
collections, or a character vector containing a file path to a data.frame}
}
\description{
This is a constructor for the Collections class. It creates a Collections
object for use in a Microbiome Dataset. A Collections object is a list of
Collection objects.
}
