% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getMetadataVariableNames}
\alias{getMetadataVariableNames}
\title{Get Microbiome Dataset Metadata Variable Names}
\usage{
getMetadataVariableNames(object)
}
\arguments{
\item{object}{A Microbiome Dataset}
}
\value{
a character vector of metadata variable names
}
\description{
Get the names of the metadata variables in the Microbiome Dataset.
}
