% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R, R/methods-MbioDataset.R
\name{getMetadataVariableNames,Collection-method}
\alias{getMetadataVariableNames,Collection-method}
\alias{getMetadataVariableNames,MbioDataset-method}
\title{Get Variable Names of Metadata}
\usage{
\S4method{getMetadataVariableNames}{Collection}(object)

\S4method{getMetadataVariableNames}{MbioDataset}(object)
}
\arguments{
\item{object}{Any object with a slot containing sampleMetadata}
}
\value{
a character vector of metadata variable names
}
\description{
Get the names of the metadata variables in an object containing sample metadata.
}
