% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{getSampleMetadataIdColumns}
\alias{getSampleMetadataIdColumns}
\title{Get Sample Metadata Id Column Names}
\usage{
getSampleMetadataIdColumns(object)
}
\arguments{
\item{object}{A Microbiome Dataset, or other object w sample metadata}
}
\value{
a character vector of id column names
}
\description{
Get the names of the record and ancestor id columns in the sample metadata of the Microbiome Dataset.
}
