% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PretermInfantResistome2}
\alias{PretermInfantResistome2}
\title{Preterm Infant Resistome 1}
\format{
\subsection{\code{PretermInfantResistome2}}{

A MbioDataset object with 24 metadata variables, 427 shotgun stool samples.
It contains the following collections: \cr
"WGS 4th level EC metagenome abundance data" \cr
"WGS Metagenomeenzyme pathway abundance data" \cr
"WGS Metagenomeenzyme pathway coverage data"  \cr
"WGS Genus" \cr
"WGS Species" \cr
"WGS Family" \cr
"WGS Order" \cr
"WGS Phylum" \cr
"WGS Class"  \cr
"WGS Kingdom"  \cr
"WGS Normalized number of taxon-specific sequence matches" \cr
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_b9dc726b20/new/download}
}
\usage{
PretermInfantResistome2
}
\description{
The Preterm Infant Resistome II (WGS) study profiled the stool metagenome in preterm infants
(born < 33 weeks gestational age) throughout the first 21 months of life.
Patients were stratified by antibiotic use during their hospitalization
41 preterm infants and 17 antibiotic-naive, healthy early term or late-preterm infants
seen at the Washington University School of Medicine in St. Louis, MO, USA
437 stool samples; 'shotgun' metagenomic sequencing.
Prospective cohort design.
}
\details{
Website Release 32. (2023 May 30)
}
\keyword{datasets}
